+++ src/main/java/ch/hutch79/command/Command.java
@@ -3,7 +3,6 @@ package ch.hutch79.command;
 import ch.hutch79.FCommand;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 
 public class Command implements CommandExecutor {
